refactor(modify): 移除圆弧切线方向计算相关方法
- 删除了 calculateArcTangentDirection 方法及其相关实现
- 移除了 isArcClockwise 方法用于判断圆弧方向的功能
- 清理了 calculateArcFallbackDirection 备用方向计算方法
- 保留了优化后的查找延伸目标点功能

+++ src/main/java/com/masterplanner/core/geometry/shapes/ArcShape.java
@@ -1393,8 +1393,8 @@ public class ArcShape extends Shape implements IExtendableShape {
                 // 从样式获取线条颜色
                 int lineColor = getStyle() != null && getStyle().getLineStyle() != null ? 
                     getStyle().getLineStyle().getColor().getRGB() : 0x80FFFFFF;
-                float lineWidth = getStyle() != null && getStyle().getLineStyle() != null ? 
-                    (float) getStyle().getLineStyle().getWidth() : 1.0f;
+                float lineWidth = getStyle() != null && getStyle().getLineStyle() != null ?
+                        getStyle().getLineStyle().getWidth() : 1.0f;
                 drawList.addLine(
                     (float) p1.x, (float) p1.y,
                     (float) p2.x, (float) p2.y,
